finish permionis

+++ Controllers/ExpensesController.cs
@@ -1,47 +1,57 @@
 ﻿using IOMSYS.IServices;
 using IOMSYS.Models;
+using IOMSYS.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 
 namespace IOMSYS.Controllers
 {
-    [Authorize(Roles = "GenralManager,BranchManager,Employee")]
+    [Authorize]
     public class ExpensesController : Controller
     {
         private readonly IExpensesService _expensesService;
         private readonly IPaymentTransactionService _paymentTransactionService;
+        private readonly IPermissionsService _permissionsService;
 
-        public ExpensesController(IExpensesService expensesService, IPaymentTransactionService paymentTransactionService)
+        public ExpensesController(IExpensesService expensesService, IPaymentTransactionService paymentTransactionService, IPermissionsService permissionsService)
         {
             _expensesService = expensesService;
             _paymentTransactionService = paymentTransactionService;
+            _permissionsService = permissionsService;
         }
 
-        [Authorize(Roles = "GenralManager,BranchManager")]
-        public IActionResult ExpensesPage()
+        public async Task<IActionResult> ExpensesPage()
         {
+            int userId = Convert.ToInt32(User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value);
+            var hasPermission = await _permissionsService.HasPermissionAsync(userId, "Expenses", "ExpensesPage");
+            if (!hasPermission) { return RedirectToAction("AccessDenied", "Access"); }
             return View();
         }
 
         [HttpGet]
-        [Authorize(Roles = "GenralManager,BranchManager,Employee")]
         public async Task<IActionResult> LoadExpenses()
         {
             var expenses = await _expensesService.GetAllExpensesAsync();
             return Json(expenses);
         }
 
-        [HttpPost]
-        [Authorize(Roles = "GenralManager,BranchManager")]
-        public async Task<IActionResult> AddNewExpense([FromForm] IFormCollection formData)
+        [HttpGet]
+        public async Task<IActionResult> LoadExpensesByBranch(int branchId)
         {
+            var Products = await _expensesService.GetAllExpensesByBranchAsync(branchId);
+            return Json(Products);
+        }
+
+        [HttpPost]
+        public async Task<IActionResult> AddNewExpense([FromBody] ExpenseModel newExpense)
+        {
+            int userId = Convert.ToInt32(User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value);
+            var hasPermission = await _permissionsService.HasPermissionAsync(userId, "Expenses", "AddNewExpense");
+            if (!hasPermission) { return BadRequest(new { ErrorMessage = "ليس لديك صلاحية" }); }
+
             try
             {
-                var values = formData["values"];
-                var newExpense = new ExpenseModel();
-                JsonConvert.PopulateObject(values, newExpense);
-
                 newExpense.PurchaseDate = DateTime.Now;
                 newExpense.UserId = Convert.ToInt32(User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value);
 
@@ -78,19 +88,22 @@ namespace IOMSYS.Controllers
         }
 
         [HttpPut]
-        [Authorize(Roles = "GenralManager,BranchManager")]
-        public async Task<IActionResult> UpdateExpense([FromForm] IFormCollection formData)
+        public async Task<IActionResult> UpdateExpense([FromBody] ExpenseModel expense)
         {
+            int userId = Convert.ToInt32(User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value);
+            var hasPermission = await _permissionsService.HasPermissionAsync(userId, "Expenses", "UpdateExpense");
+            if (!hasPermission) { return BadRequest(new { ErrorMessage = "ليس لديك صلاحية" }); }
+
             try
             {
-                var key = Convert.ToInt32(formData["key"]);
-                var values = formData["values"];
-                var expense = await _expensesService.SelectExpenseByIdAsync(key);
+                //var key = Convert.ToInt32(formData["key"]);
+                //var values = formData["values"];
+                //var expense = await _expensesService.SelectExpenseByIdAsync(key);
 
-                if (expense == null)
-                    return NotFound(new { ErrorMessage = "Expense Not Found" });
+                //if (expense == null)
+                //    return NotFound(new { ErrorMessage = "Expense Not Found" });
 
-                JsonConvert.PopulateObject(values, expense);
+                //JsonConvert.PopulateObject(values, expense);
 
                 expense.PurchaseDate = DateTime.Now;
                 expense.UserId = Convert.ToInt32(User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value);
@@ -134,12 +147,15 @@ namespace IOMSYS.Controllers
         }
 
         [HttpDelete]
-        [Authorize(Roles = "GenralManager")]
-        public async Task<IActionResult> DeleteExpense([FromForm] IFormCollection formData)
+        public async Task<IActionResult> DeleteExpense(int key)
         {
+            int userId = Convert.ToInt32(User.Claims.FirstOrDefault(c => c.Type == "UserId")?.Value);
+            var hasPermission = await _permissionsService.HasPermissionAsync(userId, "Expenses", "DeleteExpense");
+            if (!hasPermission) { return BadRequest(new { ErrorMessage = "ليس لديك صلاحية" }); }
+
             try
             {
-                var key = Convert.ToInt32(formData["key"]);
+                //var key = Convert.ToInt32(formData["key"]);
                 var expense = await _expensesService.SelectExpenseByIdAsync(key);
                 int deleteExpenseResult = await _expensesService.DeleteExpenseAsync(key);
                 if (deleteExpenseResult > 0)
